requirements.txt
python-dotenv==1.0.1
deepgram-sdk==3.4.0
elevenlabs==1.9.0
groq==0.4.1